refactor: modularize code generator
